+++ Goblin/Bot/Commands/ScheduleCommand.cs
@@ -22,7 +22,7 @@ namespace Goblin.Bot.Commands
         public string Message { get; set; }
         public Keyboard Keyboard { get; set; }
 
-        public async Task Execute(string param, int id = 0)
+        public async Task Execute(string param, long id = 0)
         {
             var user = await DbHelper.Db.Users.FirstAsync(x => x.Vk == id);
             DateTime time;
@@ -43,7 +43,7 @@ namespace Goblin.Bot.Commands
             Message = await StudentsSchedule.GetScheduleAtDate(time, user.Group);
         }
 
-        public bool CanExecute(string param, int id = 0)
+        public bool CanExecute(string param, long id = 0)
         {
             var user = DbHelper.Db.Users.First(x => x.Vk == id);
             if (user.Group == 0)
