+++ Goblin/Bot/Commands/SetGroupCommand.cs
@@ -19,7 +19,7 @@ namespace Goblin.Bot.Commands
         public string Message { get; set; }
         public Keyboard Keyboard { get; set; }
 
-        public async Task Execute(string param, int id = 0)
+        public async Task Execute(string param, long id = 0)
         {
             var group = int.Parse(param);
             var gr = StudentsSchedule.GetGroupByRealId(group);
@@ -30,7 +30,7 @@ namespace Goblin.Bot.Commands
             Message = $"Группа успешно установлена на {group} ({gr.Name})!";
         }
 
-        public bool CanExecute(string param, int id = 0)
+        public bool CanExecute(string param, long id = 0)
         {
             if (int.TryParse(param, out var i) && StudentsSchedule.IsCorrectGroup(i))
             {
