+++ Goblin/Bot/Commands/TeacherScheduleCommand.cs
@@ -18,12 +18,12 @@ namespace Goblin.Bot.Commands
         public string Message { get; set; }
         public Keyboard Keyboard { get; set; }
 
-        public async Task Execute(string param, int id = 0)
+        public async Task Execute(string param, long id = 0)
         {
             Message = await TeachersSchedule.GetScheduleToSend(int.Parse(param));
         }
 
-        public bool CanExecute(string param, int id = 0)
+        public bool CanExecute(string param, long id = 0)
         {
             if (int.TryParse(param, out var res))
             {
